Update plan for Distibuted DNS
Initial work to modify the external-dns plan to work in the proposed way
for distubuted dns.

Updates the plan to treat all endpoints passed to it (current/desired)
as part of a single shared set of records. The calculation logic now
takes into account other endpoints added to the plan when calculating
the desired changes for any other endpoint also added.

The plan calculation logic is broken into two steps:

1. Calculate what endpoints will be created/updated and deleted based on
   the current and desired records passed to the plan. For each endpoint
that will still exist if these changes were to be
applied(create/update), a map of dnsNames and their owners is
calculated.

2. Create the external-dns Changes for the endpoints calculated in
   step one. For all endpoints calculated to be an update the target
values are re-calculated using the information gathered in step one
about ownership of dnsNames. If we detect that a CNAME has a "managed"
target value (it's in our owner map) we use the information known about
who owns it to manipulate the desired targets values.