plan: Add basic conflict reporting
Update the plan to keep track of conflicts it can't resolve. It stores
each conflict as an error in an array that can be queried later before
applying changes. The plan will currently create a conflict error if the
record type of an endpoint is going to change, or an attempt to update a
record that is owned or unowned by one that is unowned or owned (The
opposite).