feat: Add providers startup flag
Add a new flag to the controller startup that determines what providers
are enabled, if none are specified the list of registered default
providers is used.

The provider factory is updated to take a list of providers that are
enabled and ensures that only providers in this list can be loaded via
the factory methods. Providers themselves optionally register themselves
as "default" providers, in the case where no providers are specified
these providers are enabled in the factory.