feat: Generate unique OwnerID based on DNSRecord UID
Allows the ownerID to initially be unset on DNSRecord creation, and the
dns record controller itself to generate and set a unique value based on
the DNSRecord UID. The ownerID is still immutable after it is set and
the dns record controller will not reconcile fully without it being set
to a valid value.

Adds additional validations to the ownerID field specifying a MinLength
and MaxLength as per best practices for CEL validations. MaxLength is
currently 36 which allows the full UID value to be used if desirable.