fix: Multiple same host/target records
Fixes an issue where if you had multiple records with the same host and
target values all contributing to the same record set, removing one of
them would cause invalid records (0 target values) to be returned from
the plan.

Fixes an issue in the inmemory provider where it was allowing records
with 0 target values to be added, and improves checks to avoid a panic.

Signed-off-by: Michael Nairn <mnairn@redhat.com>