ci: Add "CI Required Checks" job
Adds a new job to the CI jobs workflow that can be used to check the
status of all ci jobs in the workflow.

Intention is this will be used as part of the merge queue configuration
allowing a single job to be specified, inspired by
https://github.com/orgs/community/discussions/75568#discussioncomment-10351973

Add workflow_dispatch trigger to CI workflow.

Signed-off-by: Michael Nairn <mnairn@redhat.com>