ci: Add docs ci github action
Adds a check to see if we are only making docs changes when triggered
via a pull request. If we are, it will update the status of the merge
queue required checks ("CI-E2E Required Checks" and "CI Required
Checks") to prevent doc only PRs form being blocked.

Docs related jobs can be added here as needed.

Signed-off-by: Michael Nairn <mnairn@redhat.com>