feat: Add basic CoreDNS setup configuration
Adds a basic setup of CoreDNS using the upstream helm chart
https://github.com/coredns/helm wrapped in a kustomization.  The
kustomization (config/coredns) will install a CoreDNS in the
"kuadrant-dns" namespace configured with a test zone (foo.bar.baz) that
reads records from a configured db file and a default zone (.) that
forwards on all other requests to the default resolver.

Adds metallb setup to allow access to the CoreDNS instance from the host
if the host allows it.  Add grafana dashboard for CoreDNS
https://grafana.com/grafana/dashboards/14981-coredns/.

Signed-off-by: Michael Nairn <mnairn@redhat.com>