coredns(plugin): Add Zone wrapper with geo/weight answer parsing
Adds a wrapper around the file.Zone resource to deal with the refresh
and insertion(zone.Insert) of RRs in the wrapped zone from
v1alpha1.DNSRecord resources and the processing of answers received back
during lookup(zone.Lookup).

Adds basic functionality to parse weighted or geo answers in order to
reduce these down to a single answer, currently just returns the first
RR found in all cases.

Includes a change to use a forked version of coredns with updates to the
the lookup logic of the file plugin
https://github.com/mikenairn/coredns/tree/no_cache_external_lookup

Signed-off-by: Michael Nairn <mnairn@redhat.com>