coredns(plugin): Fix zone lookup
The lookup logic in
https://github.com/Kuadrant/dns-operator/commit/35d821cbfa14853f11691d6095dd15f97663db0e
was flawed as the file.Zone.lookup was always only querying the first
CNAME target when it reached an element in the cache that had multiple
options (same dns name).

The changes here add the logic to reduce the records down to a single
answer based on weight/geo into the zone lookup logic directly instead
of trying to process them after which would never work.

Also adds a temporary round robin shuffle of all weight and geo answers.
Will be removed as we progress the weight and geo resolving  based on RR
data pulled from endpoint resources.

Requires a new change to the coreDNS file plugin
https://github.com/mikenairn/coredns/tree/add_rrresolver

Signed-off-by: Michael Nairn <mnairn@redhat.com>