coredns(plugin): Implement weighted resource records
Implements `parseWeightedAnswers` based on how the default coredns
plugin works
https://github.com/coredns/coredns/blob/master/plugin/loadbalance/weighted.go.

Requests for records that are weighted will be processed to return the
most suitable based on the combined weights of the set.

Note: We can't use that plugin directly since it only deals with
weighted A records and expects a weight value to be assigned to the ip
address. We want deal with weights being assigned to any type of record
(CNAME or A), and dealt with at each stage of the recursive resolve of
CNAME records down to A.

Signed-off-by: Michael Nairn <mnairn@redhat.com>