coredns(plugin): Add watch namespaces configuration
Updates the logic in the k8s controller to setup zone informers based on
the namespaces that are being configured via the "WATCH_NAMESPACES" env
var. This behaviour is the same as the dns operator and will default to
all namespaces if the env var is not set or empty.

Signed-off-by: Michael Nairn <mnairn@redhat.com>