coredns(install): traffic policy and geoip config
Set the externalTrafficPolicy to Local in teh coredns LocdBalancer
services to preserver the client source ip.
https://kubernetes.io/docs/tasks/access-application-cluster/create-external-load-balancer/#preserving-the-client-source-ip

Add the `edns-subnet` config to the geoip plugin to use the subnet
option if it's available instead of the source ip making things easier
to test locally.

```
dig @172.18.0.17 api.k.example.com +subnet=127.0.200.200 +short
```

Signed-off-by: Michael Nairn <mnairn@redhat.com>