coredns(plugin): Limit plugins enabled in kuadrant CoreDNS
Limit the number of plugins included by default in the kuadrant build of
CoreDNS. The plugins enabled are the required ones and more simple
plugins with few additional dependencies but are useful for
debug/testing purposes.

The `trace` plugin had licensing issues and was the main reason to
reduce the plugin imports.

Signed-off-by: Michael Nairn <mnairn@redhat.com>