coredns(provider): Fix zone record retrieval
Removes the RecordsForHost() method from the provider and re-implements
coreds Records() with the same functionality.

The point of the provider.Records() method is to retrieve all the
records as they are currently in the external dns provider. In the case
of the coredns implementation that is the coredns servers themselves,
the "merged and "local" DNSRecords being created are not really relevant
here, these are just internal pieces used to achieve the update in the
provider.

Also note, the current implementation of Records() was just retrieveing
the merged record from the currently configured cluster. This won't work
when executing the e2e tests against multiple clusters as it would need
to be configured with the correct cluster context.

The CoreDNS provider is slighty different from others in that it needs
to know the exact host(DNSRecord.spec.RootHost) it's dealing with, and
not just the zone domain, an additional filter was added to the provider
config to allow this to be passed in.

A records with the same dnsname should be merged.

Signed-off-by: Michael Nairn <mnairn@redhat.com>