coredns(provider): Fix e2e tests
Update all tests to correctly check the existence of zone endpoints
based on wether or not the txt registry is enabled for a provider or
not. Disabled for CoreDNS provider.

Update checks that do a dns lookup to check all available DNS servers
to ensure the correct one is queried. In the case of coredns we don't
know for sure which one should be used, so we just have try try them all
until one or none work.

Signed-off-by: Michael Nairn <mnairn@redhat.com>