chore: Create operator resources during local-setup
Updates the local-setup logic to always apply operator resources (rbac,
service accounts etc..) from the default kustomization to all clusters.
When DEPLOY=false and DEPLOYMENT_SCOPE=cluster (defaults), the manager
Deployment and metrics Service resources are removed to allow local
development to continue as normal using `make run`.

Note: This is required to support future work where we need access to an
sa in a known namespace for local development as it would be in
production.

Signed-off-by: Michael Nairn <mnairn@redhat.com>