ci: Negate input values used to select providers
Use skip<provider> with a default value of false instead of
test<provider> with a default of true.

Inputs are only set when the job is triggered manually using the
"workflow_dispatch" event. The default needs to be that it runs the step
if this does not exist or exists and is false.

Signed-off-by: Michael Nairn <mnairn@redhat.com>