fix: status update on delete events
Fixes an issue where the status update on delete to indicate that the
finalizer can be removed was potentially not triggering an immediate
requeue of the record in order for it to remove the finalizer.

The current logic in `updateStatus` might prevent it from actually doing
any update if it decides it shouldn't, which is not compatible with the
needs of the status update for deletion events.

Signed-off-by: Michael Nairn <mnairn@redhat.com>