Add delegation role controller option
Adds a controller option "--delegation-role=[primary|secondary]" that
depending on the value will initiate the controller in different "modes".

A role of "secondary" indicates that it will have limited
responsibilities when reconciling DNSRecords that are delegated, in this
mode the controller will not start the multi cluster reocnciler and
cluster secrets will not be processed.A role of "primary" is the default
and indicates it should start with multicluster features enabled(current
behaviour) and will process delegated records (future work).

Signed-off-by: Michael Nairn <mnairn@redhat.com>