add ready for delegation status condition
Adds a new status condition to the DNSRecord resource
"ReadyForDelegation" that indicates whether or not a delegating record
has been processed by the local dns operator instance. Any remote
reconcilers that see this record should ignore it until this status
exists and is set to true.

Note: Is currently set to true only when the finalizer has been added to
the DNSRecord resource, this will need to be updated to also take into
account health check status.

Signed-off-by: Michael Nairn <mnairn@redhat.com>