test: Update integration test suite for primary and secondary testing
Updates the integration test suite to make it possible to test both
primary and secondary functionality in isolation or together in more
comprehensive test cases.

The test suite creates two controller-runtime test environments (k8s api
server, etcd etc..) on initialisation, one to act as a primry and one as
a secondary. Each test environment runs the appropriate components for
the given delegation role(based on the main application):

Primary:
   - create multicluster-controller-runtime manager
   - setup kubeconfig provider
   - setup DNSRecordReconciler
   - setup RemoteDNSRecordReconciler
Secondary:
   - create controller-runtime manager
   - setup DNSRecordReconciler

The test suite contains a global reference to the primary and secondary
environments, managers, k8s clients and kubeconfigs for a dedicated user
"kuadrant" allowing them to be used in specs as required.

Signed-off-by: Michael Nairn <mnairn@redhat.com>