test: Add kubeconfig provider tests
Adds basic tests to the integration test suite to validate the behaviour
of the kubeconfig provider for multicluster secrets.

Makes use of `gbytes.Buffer` to capture log statements during the
current test and the gomega `gbytes.Say` matcher see
https://onsi.github.io/gomega/#codegbytescode-testing-streaming-buffers
since there is no way of really checking the status success/failure of
these secrets from any particular resource. Logs in the test suite are
pre-pended with the delegation role (primary or secondary) so it can be
determined which environment is logging the statements.

Signed-off-by: Michael Nairn <mnairn@redhat.com>