fix remote dnsrecord controller logging
Fixes an issue with inconsistent logging between a DNSRecord
reconciliation triggered by a local or remote event.

The RemoteDNSRecordReconciler creates a request specific
DNSRecordReconciler instance, and is not setup via the manager meaning
the log constructor that would normally be used is not called. The
changes here build up the same logger and set it in the context for
consistency in log statements.

Disable dev mode logging in integration tests and query expected
structure logs in tests.

Signed-off-by: Michael Nairn <mnairn@redhat.com>