fix: use hashed root host value for auth record labels
The dnsrecord root host value was being used for the value of the
delegation auth record label
("kuadrant.io/delegation-authoritative-record") which would fail if the
root host value was longer than the maximum allowed length of a label
value (63 characters). The change here modifies this to instead use a
hash with a fixed length of eight of the root host instead.

Signed-off-by: Michael Nairn <mnairn@redhat.com>