apply ready for delegation status to all delegating records
Updates the reconciliation logic to apply the "ReadyForDelegation"
status to all delegating DNSRecords regardless of it being processed as
a primary or secondary.

In the case of "primry-primary" setups the "ReadyForDelegation" status
will need to exist on all delegating records in order for another
primary to process it as a remote record that should be included in its
own authoritative record.

Signed-off-by: Michael Nairn <mnairn@redhat.com>