update delegation authoritative record naming
Updates the naming of the authoritative dnsrecord created by the
delegation feature to base the name on a hash of the root host rather
than generating a random name. Naming the auth records in this way
allows them to be deterministic, and in environments with more than one
primary allows the names to match across clusters conforming to the
sameness principals of k8s multi cluster that we want to achieve.

Signed-off-by: Michael Nairn <mnairn@redhat.com>