refactor: delegation and authoritative record creation logic
Moves the delegation code into the provider package. The provider itself
is now adapted for delegating instances to ensure the existence of the
required authoritative record when any requests to `DNSZoneForHost` are
made. Also adds the ability for individual providers to require
delegation implicitly if they required it, this will be the case for the
CoreDNS provider in upcoming changes.

Signed-off-by: Michael Nairn <mnairn@redhat.com>