refactor: delegation and authoritative record creation logic
Moves the delegation logic into the provider package. The provider
factory is now configured with a dedicated delegation provider, calls to
`factory.ProviderFor` will inspect the accessor and return an instance
of this provider when delegation is required.

Adds a new internal provider `AuthoritativeDNSRecordProvider` composed
of the endpoint provider that modifies behaviour specifically for the
management of authoritative DNSRecord resources for the delegation
feature. Calls to `DNSZoneForHost` will ensure the existence of the
expected authoritative record before calling the normal endpoint impl.
All instances of the provider factory are updated to use the
`AuthoritativeDNSRecordProvider` for delegation.

Adds the ability for individual providers to require delegation
implicitly if they required it, this will be the case for the CoreDNS
provider in upcoming changes.

Signed-off-by: Michael Nairn <mnairn@redhat.com>