RemoteDNSRecordReconciler: update apply changes
Remove health check manipulation of endpoints
Remove zoneEndpoints being set in the status

I don't believe either of these things, as well as status updates,
should be done in here for either reconciler. Holding off on a
changing the local one for now, can be dealt with later when we have a
better idea of the commonality of the two.

DNSRecordReconciler: removes extra health probe check

Signed-off-by: Michael Nairn <mnairn@redhat.com>