feat: Add support for multiple primary controllers
Adds a new field to the status of the DNSRecord resource
`remoteRecordStatuses` which is intended to hold primary cluster
specific statuses, one for each primary, that is reconciling it as a
"remote" and adding it's endpoints to a local authoritative record.
A cluster UID value is generated from the kube-system namespace for a
repeatable cluster specific ID, and used as the key in the
`remoteRecordStatuses` map, with each entry being a `DNSRecordStatus`
for that cluster.

The RemoteDNSRecordReconciler is updated to reconcile the remote record
by loading and updating the correct cluster specific status in
`remoteRecordStatuses` instead of the normal `record.Status`. A
`RemoteDNSRecord` wrapper and `DNSRecordAccessor` interface were added
to help achieve the required loading and setting of the status, with the
intention of making this a pattern used in other reconcilers at a later
date to reduce the amount of duplicated code.

Most changes are confined to the remote record reconciliation logic, but
some changes were made to the checks in `ProviderEndpointsRemoved` to
ensure during deletion events that all remote clusters have the chance to
clean up before the dnsrecord finalizer is removed.

Signed-off-by: Michael Nairn <mnairn@redhat.com>