provider(coredns): remove NAMESERVERS from config
NAMESERVERS is no longer required by the CoreDNS provider, this removes
it as a mandatory piece of data in the CoreDNS provider secret.

Note: Our e2e test suite is currently using the NAMESERVERS data in the
secret so it is still being generated and added as part of local
setup. It is not required for the CoreDNS provider to function though,
and will look to remove it completely at a later stage.

Signed-off-by: Michael Nairn <mnairn@redhat.com>