provider(authoritative_dnsecord): Manage auth record lifecycle
Updates the AuthoritativeDNSRecordProvider to better manage the
lifecycle(create/update/delete) of the authoritative dnsrecord being
used for the delegating resource.

The provider now ensures the existence of the authoritative record on
calls into `ApplyChanges`, creating the record if required, before
applying changes.  If the delegating record is deleting the
authoritative record will also be removed if the result of applying
changes is an empty record.
The expected labels on the authoritative record i.e.
`kuadrant.io/authoritative-record=true` and
`kuadrant.io/authoritative-record-hash=xyz`  will be added when applying
changes if they do not exist.

Internally `getAuthoritativeRecord` was also updated to get the record
by name i.e. `authoritative-record-<rootHostHash>` rather than listing
records to match how `createAuthoritativeRecord` works.

Signed-off-by: Michael Nairn <mnairn@redhat.com>