fix(iteration): Add runtime group to record status
Updates Registry.AdjustEndpoints to add group and target labels to all
endpoints if a group is set in the record status.

* Add `GetGroup()` to `DNSRecordAccessor` interface
* Add group to logger
* Set status group in local record reconcile

Health checks:

Move removeUnhealthyEndpoints call to `getEndpoints()` to get round an
issue where the spec endpoints are being incorrectly modified and saved
to the spec when saving the provider labels.

Rename external dns imports:

Rename all external-dns imports, all imports that are coming directly
from the external dns module are prepended with "externaldns", internal
package versions do not.

Signed-off-by: Michael Nairn <mnairn@redhat.com>