Refactor: list of objects to list of strings
Change the list of Alpha 2 country code objects to a list of strings. All
access to the list was done via a function that converted the list of
objects into a list of strings. This function has now being removed.

The custom ContainsString function has also being removed in favor of
the standard libraries slices.Contains function.

There is no change in behaviour, but the changes should reduce the number
of cpu cycles required to get the same behaviour.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>